Infer source families and refine aggregate typing

+++ decomp/ST.exe/functions/00483F10/decomp.c
@@ -18,7 +18,8 @@ undefined4 __fastcall FUN_00483f10(STBoatC *param_1)
   longlong lVar9;
   longlong lVar10;
 
-  iVar3 = STAllPlayersC::IsTorpLive(g_allPlayers_007FA174,param_1->field_046F,param_1->field_0471);
+  iVar3 = STAllPlayersC::IsTorpLive
+                    (g_allPlayers_007FA174,param_1->field_046F,(int)param_1->field_0471);
   if (iVar3 != 0) {
     puVar1 = param_1->field_046B;
     uVar4 = (int)param_1->field_0045 - (int)*(short *)((int)puVar1 + 0x45);
